documentation and tasks

+++ Naped_DC_PV_SC/include/sensors.h
@@ -1,6 +1,6 @@
 #ifndef SENSORS_H
 #define SENSORS_H
 
-float read_current(int, int);
+int read_current(int, int);
 
 #endif
